feat: add skad_report table and remove duplicated code
