+++ adjust_report_etl/.gitignore
@@ -1,0 +1,3 @@
+__pycache__/
+myenv/
+files/
