--- adjust_report_etl/Feast/configuration.json
@@ -1,4 +1,0 @@
-{
-  "API_KEY": "xxx",
-  "APP_TOKEN": "yyy"
-}
