+++ adjust_report_etl/configuration/adjust_premium.json
@@ -1,0 +1,4 @@
+{
+    "API_KEY": "xxx",
+    "APP_TOKEN": "xxx"
+}
